refactor: Food Repository 패키지명 변경 - meal -> food로 패키지명 변경
Issue: #35
